make sure address is really set

+++ src/entrypoint/commands/complete_multipart.cpp
@@ -118,7 +118,7 @@ coro<response> complete_multipart::handle(request& req) {
     co_await m_directory.put_object(req.bucket(), obj);
 
     if constexpr (m_enable_refcount) {
-        if (old_obj.has_value()) {
+        if (old_obj.has_value() && old_obj->addr.has_value()) {
             co_await m_gdv.unlink(req.context(), old_obj.value().addr.value());
         }
     }
